requirements.txt
anyio
certifi
click
fastapi
h11
idna
Jinja2
MarkupSafe
psycopg2-binary
pydantic
PyJWT
python-dotenv
python-multipart
setuptools
six
sniffio
starlette
typing_extensions
uvicorn
Werkzeug
wheel
